Use fetcher for fetching resources
The fetecher will log you in, and will maintain the necessary session
cookie in order to remember who you are.

Future work: allow consumers to 'reset' the login or connect services so
we know we have to login again.  This may happen when an identifier
on the other stack gets a new identifier secret or when the database is
instructed to remove our session.

Pairing with @erikap