feat: invert deletes and inserts and execute them by graph
moveTriples now takes into account the graph of the quads and executes
the delete before the insert, because some producers store ALL the
deltas something like a PATCH will contain an insert of a triple and a
delete of the old triple, but given that Ember Data patches update all
the data, the delete will include unchanged triples. If we execute the
delete last we will lose data.

To prevent many inserts/deletes we group the quads by their graph, so
that we only execute 1 insert/delete per graph.