docs: update the README a little bit
The README still needs a proper rewrite but a lot of what's in it is
still based on the tripled-dispatching behaviour which this branch seems
to want to rewrite, but that rewrite isn't fully finished either given
that the initial sync pipeline still uses the old code.

If this branch is something that we want to merge into master at some
point (instead of only using it for Kaleidos' server mirroring) we
should look into finishing it.