Improve failure of job.
Previously if a job was not finished it failed all tasks. Even if the
task was a success. Since we use the last success task to calculate
the next delta to consume; this might lead restart too many tasks
being redone again.
By failing only the non success tasks; when we restart we restart
right at where we left.